General: moved styles to assets/jss
