+++ src/assets/jss/contactPanelStyle.js
@@ -1,0 +1,13 @@
+import amber from '@material-ui/core/colors/amber';
+
+const contactPanelStyle = theme => ({
+  root: {
+    width: '100%',
+    backgroundColor: theme.palette.background.paper,
+  },
+  star: {
+    color: amber[500]
+  }
+});
+
+export default contactPanelStyle;
