+++ src/assets/jss/expansionPanelStyle.js
@@ -1,0 +1,10 @@
+const expansionPanelStyle = theme => ({
+  heading: {
+    fontWeight: theme.typography.fontWeightMedium,
+  },
+  disablePadding: {
+    padding: "0px"
+  }
+});
+
+export default expansionPanelStyle;
