+++ src/assets/jss/tableStyle.js
@@ -1,0 +1,8 @@
+const tableStyle = theme => ({
+  tableCell : {
+    paddingLeft: '18px',
+    paddingRight: '0px'
+  }
+});
+
+export default tableStyle;
