+++ src/views/InfoView/panels/CantusPanel.js
@@ -5,7 +5,7 @@ import Typography from '@material-ui/core/Typography';
 
 import DefaultPanel from '../DefaultPanel';
 
-import infoPanelStyle from '../../../assets/jss/infoPanelStyle';
+import boldStyle from '../../../assets/jss/boldStyle';
 
 
 function CantusPanel(props) {
@@ -29,4 +29,4 @@ CantusPanel.propTypes = {
   classes: PropTypes.object.isRequired,
 };
 
-export default withStyles(infoPanelStyle)(CantusPanel);
+export default withStyles(boldStyle)(CantusPanel);
