+++ src/views/InfoView/panels/SubrentPanel.js
@@ -5,7 +5,7 @@ import Typography from '@material-ui/core/Typography';
 
 import DefaultPanel from '../DefaultPanel';
 
-import infoPanelStyle from '../../../assets/jss/infoPanelStyle';
+import boldStyle from '../../../assets/jss/boldStyle';
 
 
 function SubrentPanel(props) {
@@ -29,4 +29,4 @@ SubrentPanel.propTypes = {
   classes: PropTypes.object.isRequired,
 };
 
-export default withStyles(infoPanelStyle)(SubrentPanel);
+export default withStyles(boldStyle)(SubrentPanel);
